+++ e2e/__tests__/coverageHandlebars.test.ts
@@ -8,7 +8,7 @@
 import * as path from 'path';
 import {readFileSync} from 'graceful-fs';
 import wrap from 'jest-snapshot-serializer-raw';
-import {cleanup, run} from '../Utils';
+import {cleanup, runYarn} from '../Utils';
 import runJest from '../runJest';
 
 const dir = path.resolve(__dirname, '../coverage-handlebars');
@@ -19,7 +19,7 @@ beforeAll(() => {
 });
 
 it('code coverage for Handlebars', () => {
-  run('yarn', dir);
+  runYarn(dir);
   const result = runJest(dir, ['--coverage', '--no-cache']);
 
   expect(result.exitCode).toBe(0);