+++ e2e/__tests__/typescriptCoverage.test.ts
@@ -7,12 +7,12 @@
 
 import * as path from 'path';
 import {wrap} from 'jest-snapshot-serializer-raw';
-import {run} from '../Utils';
+import {runYarn} from '../Utils';
 import runJest from '../runJest';
 
 it('instruments and collects coverage for typescript files', () => {
   const dir = path.resolve(__dirname, '../typescript-coverage');
-  run('yarn', dir);
+  runYarn(dir);
   const {stdout} = runJest(dir, ['--coverage', '--no-cache'], {
     stripAnsi: true,
   });