+++ packages/jest-jasmine2/src/jasmineAsyncInstall.ts
@@ -17,26 +17,30 @@ import throat from 'throat';
 import isError from './isError';
 import type {Jasmine} from './types';
 import type Spec from './jasmine/Spec';
-import type {QueueableFn} from './queueRunner';
+import type {DoneFn, QueueableFn} from './queueRunner';
 
-interface DoneFn {
-  (): void;
-  fail: (error: Error) => void;
-}
-
-function isPromise(obj: any) {
+function isPromise(obj: any): obj is PromiseLike<unknown> {
   return obj && typeof obj.then === 'function';
 }
 
+const doneFnNoop = () => {};
+
+doneFnNoop.fail = () => {};
+
 function promisifyLifeCycleFunction(
-  originalFn: Function,
+  originalFn: (beforeAllFunction: QueueableFn['fn'], timeout?: number) => void,
   env: Jasmine['currentEnv_'],
 ) {
   return function <T>(
-    fn: Function | (() => Promise<T>) | GeneratorFunction | undefined,
+    fn:
+      | ((done: DoneFn) => void | PromiseLike<T>)
+      | (() => Promise<T>)
+      | GeneratorFunction
+      | undefined,
     timeout?: number,
-  ) {
+  ): void {
     if (!fn) {
+      // @ts-expect-error: missing fn arg is handled by originalFn
       return originalFn.call(env);
     }
 
@@ -59,7 +63,7 @@ function promisifyLifeCycleFunction(
     // didn't return a promise.
     const asyncJestLifecycle = function (done: DoneFn) {
       const wrappedFn = isGeneratorFn(fn) ? co.wrap(fn) : fn;
-      const returnValue = wrappedFn.call({});
+      const returnValue = wrappedFn.call({}, doneFnNoop);
 
       if (isPromise(returnValue)) {
         returnValue.then(done.bind(null, null), (error: Error) => {
@@ -82,12 +86,21 @@ function promisifyLifeCycleFunction(
 // Similar to promisifyLifeCycleFunction but throws an error
 // when the return value is neither a Promise nor `undefined`
 function promisifyIt(
-  originalFn: Function,
+  originalFn: (
+    description: string,
+    fn: QueueableFn['fn'],
+    timeout?: number,
+  ) => Spec,
   env: Jasmine['currentEnv_'],
   jasmine: Jasmine,
 ) {
-  return function (specName: string, fn: Function, timeout?: number) {
+  return function (
+    specName: string,
+    fn?: (done: DoneFn) => void | PromiseLike<void>,
+    timeout?: number,
+  ): Spec {
     if (!fn) {
+      // @ts-expect-error: missing fn arg is handled by originalFn
       const spec = originalFn.call(env, specName);
       spec.pend('not implemented');
       return spec;
@@ -109,7 +122,7 @@ function promisifyIt(
 
     const asyncJestTest = function (done: DoneFn) {
       const wrappedFn = isGeneratorFn(fn) ? co.wrap(fn) : fn;
-      const returnValue = wrappedFn.call({});
+      const returnValue = wrappedFn.call({}, doneFnNoop);
 
       if (isPromise(returnValue)) {
         returnValue.then(done.bind(null, null), (error: Error) => {