+++ packages/jest-resolve/src/defaultResolver.ts
@@ -12,6 +12,7 @@ import {tryRealpath} from 'jest-util';
 import type {Config} from '@jest/types';
 
 type ResolverOptions = {
+  allowPnp?: boolean;
   basedir: Config.Path;
   browser?: boolean;
   defaultResolver: typeof defaultResolver;
@@ -21,12 +22,20 @@ type ResolverOptions = {
   rootDir?: Config.Path;
 };
 
+declare global {
+  namespace NodeJS {
+    export interface ProcessVersions {
+      // the "pnp" version named isn't in DefinitelyTyped
+      pnp?: unknown;
+    }
+  }
+}
+
 export default function defaultResolver(
   path: Config.Path,
   options: ResolverOptions,
 ): Config.Path {
-  // @ts-expect-error: the "pnp" version named isn't in DefinitelyTyped
-  if (process.versions.pnp) {
+  if (process.versions.pnp && options.allowPnp !== false) {
     return pnpResolver(path, options);
   }