+++ packages/jest-resolve/src/isBuiltinModule.ts
@@ -8,9 +8,7 @@
 import module = require('module');
 
 // "private" api
-declare const process: {
-  binding(type: string): {};
-};
+declare const process: NodeJS.Process & {binding(type: string): {}};
 
 const EXPERIMENTAL_MODULES = ['worker_threads'];