+++ packages/jest-runtime/src/helpers.ts
@@ -10,6 +10,25 @@ import slash = require('slash');
 import glob = require('glob');
 import type {Config} from '@jest/types';
 
+const OUTSIDE_JEST_VM_PROTOCOL = 'jest-main:';
+// String manipulation is easier here, fileURLToPath is only in newer Nodes,
+// plus setting non-standard protocols on URL objects is difficult.
+export const createOutsideJestVmPath = (path: string): string =>
+  OUTSIDE_JEST_VM_PROTOCOL + '//' + encodeURIComponent(path);
+export const decodePossibleOutsideJestVmPath = (
+  outsideJestVmPath: string,
+): string | undefined => {
+  if (outsideJestVmPath.startsWith(OUTSIDE_JEST_VM_PROTOCOL)) {
+    return decodeURIComponent(
+      outsideJestVmPath.replace(
+        new RegExp('^' + OUTSIDE_JEST_VM_PROTOCOL + '//'),
+        '',
+      ),
+    );
+  }
+  return undefined;
+};
+
 export const findSiblingsWithFileExtension = (
   moduleFileExtensions: Config.ProjectConfig['moduleFileExtensions'],
   from: Config.Path,