+++ scripts/cleanE2e.js
@@ -11,7 +11,7 @@ const {normalize, resolve} = require('path');
 const {sync: glob} = require('glob');
 const {sync: rimraf} = require('rimraf');
 
-const blacklist = [
+const excludedModules = [
   'e2e/global-setup-node-modules/node_modules/',
   'e2e/presets/js/node_modules/',
   'e2e/presets/json/node_modules/',
@@ -19,7 +19,7 @@ const blacklist = [
 
 const e2eNodeModules = glob('e2e/*/node_modules/')
   .concat(glob('e2e/*/*/node_modules/'))
-  .filter(dir => !blacklist.includes(dir))
+  .filter(dir => !excludedModules.includes(dir))
   .map(dir => resolve(__dirname, '..', dir))
   .sort();